Add files via upload
This should start a session and store the values there. Also it reads the value from POST  to session data and then from session data to local variables. It should then show a greeting ( or back to log in form , if login isn't successful). If login is successful , username should be displayed.